requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
pytest
selenium
